chore: remove lib/heliaFetch and cleanup sw.ts (#95)
