+++ package.json
@@ -37,7 +37,6 @@
     "@libp2p/logger": "^4.0.6",
     "@sgtpooki/file-type": "^1.0.1",
     "debug": "^4.3.4",
-    "mime-types": "^2.1.35",
     "multiformats": "^11.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
