+++ src/sw.ts
@@ -1,22 +1,53 @@
-import mime from 'mime-types'
 import { getVerifiedFetch } from './get-helia.ts'
 import { HeliaServiceWorkerCommsChannel, type ChannelMessage } from './lib/channel.ts'
 import { getSubdomainParts } from './lib/get-subdomain-parts.ts'
-import { getVerifiedFetchUrl, heliaFetch } from './lib/heliaFetch.ts'
 import { error, log, trace } from './lib/logger.ts'
 import { findOriginIsolationRedirect } from './lib/path-or-subdomain.ts'
 import type { VerifiedFetch } from '@helia/verified-fetch'
 
+/**
+ ******************************************************
+ * Types
+ ******************************************************
+ */
+
+/**
+ * Not available in ServiceWorkerGlobalScope
+ */
+interface AggregateError extends Error {
+  errors: Error[]
+}
+
+interface FetchHandlerArg {
+  path: string
+  request: Request
+
+}
+interface GetVerifiedFetchUrlOptions {
+  protocol?: string | null
+  id?: string | null
+  path: string
+}
+
+/**
+ ******************************************************
+ * "globals"
+ ******************************************************
+ */
 declare let self: ServiceWorkerGlobalScope
-
 let verifiedFetch: VerifiedFetch
+const channel = new HeliaServiceWorkerCommsChannel('SW')
+const urlInterceptRegex = [new RegExp(`${self.location.origin}/ip(n|f)s/`)]
 
+/**
+ ******************************************************
+ * Service Worker Lifecycle Events
+ ******************************************************
+ */
 self.addEventListener('install', (event) => {
   // 👇 When a new version of the SW is installed, activate immediately
   void self.skipWaiting()
 })
-
-const channel = new HeliaServiceWorkerCommsChannel('SW')
 
 self.addEventListener('activate', () => {
   // Set verified fetch initially
@@ -38,25 +69,75 @@ self.addEventListener('activate', () => {
   })
 })
 
-/**
- * Not available in ServiceWorkerGlobalScope
- */
+self.addEventListener('fetch', event => {
+  const request = event.request
+  const urlString = request.url
+  const url = new URL(urlString)
 
-interface AggregateError extends Error {
-  errors: Error[]
+  if (!isValidRequestForSW(event)) {
+    trace('helia-sw: not a valid request for helia-sw, ignoring ', urlString)
+    return
+  } else {
+    log('helia-sw: valid request for helia-sw: ', urlString)
+  }
+
+  if (isRootRequestForContent(event)) {
+    // intercept and do our own stuff...
+    event.respondWith(fetchHandler({ path: url.pathname, request }))
+  } else if (isSubdomainRequest(event)) {
+    event.respondWith(fetchHandler({ path: url.pathname, request }))
+  }
+})
+
+/**
+ ******************************************************
+ * Functions
+ ******************************************************
+ */
+function isRootRequestForContent (event: FetchEvent): boolean {
+  const urlIsPreviouslyIntercepted = urlInterceptRegex.some(regex => regex.test(event.request.url))
+  const isRootRequest = urlIsPreviouslyIntercepted
+  return isRootRequest // && getCidFromUrl(event.request.url) != null
+}
+
+function isSubdomainRequest (event: FetchEvent): boolean {
+  const { id, protocol } = getSubdomainParts(event.request.url)
+  trace('isSubdomainRequest.id: ', id)
+  trace('isSubdomainRequest.protocol: ', protocol)
+
+  return id != null && protocol != null
+}
+
+function isValidRequestForSW (event: FetchEvent): boolean {
+  return isSubdomainRequest(event) || isRootRequestForContent(event)
 }
 
 function isAggregateError (err: unknown): err is AggregateError {
   return err instanceof Error && (err as AggregateError).errors != null
 }
 
-interface FetchHandlerArg {
-  path: string
-  request: Request
+function getVerifiedFetchUrl ({ protocol, id, path }: GetVerifiedFetchUrlOptions): string {
+  if (id != null && protocol != null) {
+    return `${protocol}://${id}${path}`
+  }
 
+  const pathParts = path.split('/')
+
+  let pathPartIndex = 0
+  let namespaceString = pathParts[pathPartIndex++]
+  if (namespaceString === '') {
+  // we have a prefixed '/' in the path, use the new index instead
+    namespaceString = pathParts[pathPartIndex++]
+  }
+  if (namespaceString !== 'ipfs' && namespaceString !== 'ipns') {
+    throw new Error(`only /ipfs or /ipns namespaces supported got ${namespaceString}`)
+  }
+  const pathRootString = pathParts[pathPartIndex++]
+  const contentPath = pathParts.slice(pathPartIndex++).join('/')
+  return `${namespaceString}://${pathRootString}/${contentPath}`
 }
 
-const fetchHandler = async ({ path, request }: FetchHandlerArg): Promise<Response> => {
+async function fetchHandler ({ path, request }: FetchHandlerArg): Promise<Response> {
   // test and enforce origin isolation before anything else is executed
   const originLocation = await findOriginIsolationRedirect(new URL(request.url))
   if (originLocation !== null) {
@@ -76,13 +157,21 @@ const fetchHandler = async ({ path, request }: FetchHandlerArg): Promise<Respons
    * * https://bugzilla.mozilla.org/show_bug.cgi?id=1394102
    */
   // 5 minute timeout
-  const abortController = AbortSignal.timeout(5 * 60 * 1000)
+  const signal = AbortSignal.timeout(5 * 60 * 1000)
   try {
     const { id, protocol } = getSubdomainParts(request.url)
     const verifiedFetchUrl = getVerifiedFetchUrl({ id, protocol, path })
     log('verifiedFetch for ', verifiedFetchUrl)
 
-    return await heliaFetch({ verifiedFetch, verifiedFetchUrl, signal: abortController, headers: request.headers })
+    const headers = request.headers
+    return await verifiedFetch(verifiedFetchUrl, {
+      signal,
+      headers,
+      // TODO redirect: 'manual', // enable when http urls are supported by verified-fetch: https://github.com/ipfs-shipyard/helia-service-worker-gateway/issues/62#issuecomment-1977661456
+      onProgress: (e) => {
+        trace(`${e.type}: `, e.detail)
+      }
+    })
   } catch (err: unknown) {
     const errorMessages: string[] = []
     if (isAggregateError(err)) {
@@ -103,88 +192,3 @@ const fetchHandler = async ({ path, request }: FetchHandlerArg): Promise<Respons
     return new Response('heliaFetch error: ' + errorMessage, { status: 500 })
   }
 }
-
-const urlInterceptRegex = [new RegExp(`${self.location.origin}/ip(n|f)s/`)]
-
-/**
- *
- * @param event
- * @todo make this smarter
- * @returns
- */
-const isReferrerPreviouslyIntercepted = (event: FetchEvent): boolean => {
-  return urlInterceptRegex.some(regex => regex.test(event.request.referrer)) // && getCidFromUrl(event.request.referrer) != null
-}
-
-const isRootRequestForContent = (event: FetchEvent): boolean => {
-  const urlIsPreviouslyIntercepted = urlInterceptRegex.some(regex => regex.test(event.request.url))
-  const isRootRequest = !isReferrerPreviouslyIntercepted(event) && urlIsPreviouslyIntercepted
-  return isRootRequest // && getCidFromUrl(event.request.url) != null
-}
-
-function isSubdomainRequest (event: FetchEvent): boolean {
-  const { id, protocol } = getSubdomainParts(event.request.url)
-  trace('isSubdomainRequest.id: ', id)
-  trace('isSubdomainRequest.protocol: ', protocol)
-
-  return id != null && protocol != null
-}
-
-const isValidRequestForSW = (event: FetchEvent): boolean =>
-  isSubdomainRequest(event) || isRootRequestForContent(event) || isReferrerPreviouslyIntercepted(event)
-
-self.addEventListener('fetch', event => {
-  const request = event.request
-  const urlString = request.url
-  const url = new URL(urlString)
-
-  if (!isValidRequestForSW(event)) {
-    trace('helia-sw: not a valid request for helia-sw, ignoring ', urlString)
-    return
-  } else {
-    log('helia-sw: valid request for helia-sw: ', urlString)
-  }
-
-  if (isReferrerPreviouslyIntercepted(event)) {
-    log('helia-sw: referred from ', request.referrer)
-    const destinationParts = urlString.split('/')
-    const referrerParts = request.referrer.split('/')
-    const newParts: string[] = []
-    let index = 0
-    while (destinationParts[index] === referrerParts[index] && index < destinationParts.length && index < referrerParts.length) {
-      newParts.push(destinationParts[index])
-      index++
-    }
-    newParts.push(...referrerParts.slice(index))
-
-    const newUrlString = newParts.join('/') + '/' + destinationParts.slice(index).join('/')
-    const newUrl = new URL(newUrlString)
-
-    /**
-     * respond with redirect to newUrl
-     */
-    if (newUrl.toString() !== urlString) {
-      log('helia-sw: rerouting request to: ', newUrl.toString())
-      const redirectHeaders = new Headers()
-      redirectHeaders.set('Location', newUrl.toString())
-      if (mime.lookup(newUrl.toString()) != null) {
-        redirectHeaders.set('Content-Type', mime.lookup(newUrl.toString()))
-      }
-      redirectHeaders.set('X-helia-sw', 'redirected')
-      const redirectResponse = new Response(null, {
-        status: 308,
-        headers: redirectHeaders
-      })
-      event.respondWith(redirectResponse)
-    } else {
-      log('helia-sw: not rerouting request to same url: ', newUrl.toString())
-
-      event.respondWith(fetchHandler({ path: url.pathname, request }))
-    }
-  } else if (isRootRequestForContent(event)) {
-    // intercept and do our own stuff...
-    event.respondWith(fetchHandler({ path: url.pathname, request }))
-  } else if (isSubdomainRequest(event)) {
-    event.respondWith(fetchHandler({ path: url.pathname, request }))
-  }
-})
